♻  refactor: Rename `all` utils to `common` and move `py_ver` there
