+++ src/siesta/cli.py
@@ -31,7 +31,7 @@ from cyclopts import App
 from rich import print
 from watchdog.observers import Observer
 
-from siesta.utils.all import (
+from siesta.utils.common import (
     load_deps,
     logger,
     resolve_path,
