+++ src/siesta/utils/docs.py
@@ -12,8 +12,9 @@ from tempfile import TemporaryDirectory
 from rich import print
 from watchdog.events import FileSystemEvent, RegexMatchingEventHandler
 
-from siesta.utils.all import (
+from siesta.utils.common import (
     ROOT,
+    get_pyver,
     load_deps,
     logger,
     resolve_path,
@@ -155,26 +156,6 @@ def write_rtd_config() -> None:
 
     safe_dump(config, rtd)
     logger.info("ReadTheDocs file written.")
-
-
-def get_pyver():
-    """Get the Python version from the user.
-
-    Returns
-    -------
-    str
-        The Python version.
-    """
-    python_version_file = Path(".python-version")
-    if python_version_file.exists():
-        return python_version_file.read_text().strip()
-    if run_command(["which", "uv"]):
-        # e.g. "Python 3.12.1"
-        full_version = run_command(["uv", "run", "python", "--version"]).stdout.strip()
-        version = full_version.split()[1]
-        major, minor, _ = version.split(".")
-        return f"{major}.{minor}"
-    return "3.12"
 
 
 def update_conf_py(dest: Path, branch: str = "main"):
